feat: get user from the API at the top level in <App> and not in <UserMenu>
