+++ src/api/authentication.ts
@@ -3,6 +3,8 @@ import "firebase/auth"
 import firebase from "./firebase"
 
 const provider = new firebase.auth.GoogleAuthProvider()
-export const signInWithGoogle = firebase.auth().signInWithPopup.bind(firebase.auth(), provider)
+export const signInWithGoogle = () => firebase.auth().signInWithPopup(provider)
 
-export const signOut = firebase.auth().signOut.bind(firebase.auth())
+export const signOut = () => firebase.auth().signOut()
+
+export const onAuthStateChanged = firebase.auth().onAuthStateChanged.bind(firebase.auth())
